ProjectileShooter.cs: Inserted some code as temporary workaround for projectil-always-handdraggable bug
It might be good design that players cant mess with balls rolling around on the floor, but simple including the workaround I implemented does not address the underlying problem that the projectiles are handdraggable w/out seeming to need to be focused on. I have comment around all of the 'workaround code' in ProjectileShooter.cs with FIXME comments, so changes can be made later.